refactor(config): split static settings ports
