+++ src/main/java/me/golemcore/bot/domain/memory/retrieval/MemoryCandidateCollector.java
@@ -35,7 +35,7 @@ import me.golemcore.bot.domain.model.MemoryItem;
 import me.golemcore.bot.domain.service.MemoryScopeSupport;
 import me.golemcore.bot.domain.service.RuntimeConfigService;
 import me.golemcore.bot.port.outbound.StoragePort;
-import me.golemcore.bot.port.outbound.BotSettingsPort;
+import me.golemcore.bot.port.outbound.MemorySettingsPort;
 import org.springframework.stereotype.Service;
 
 /**
@@ -51,13 +51,13 @@ public class MemoryCandidateCollector {
     private static final String PROCEDURAL_FILE = "items/procedural.jsonl";
 
     private final StoragePort storagePort;
-    private final BotSettingsPort settingsPort;
+    private final MemorySettingsPort settingsPort;
     private final RuntimeConfigService runtimeConfigService;
     private final ObjectMapper objectMapper;
 
     public MemoryCandidateCollector(
             StoragePort storagePort,
-            BotSettingsPort settingsPort,
+            MemorySettingsPort settingsPort,
             RuntimeConfigService runtimeConfigService,
             ObjectMapper objectMapper) {
         this.storagePort = storagePort;
