+++ src/main/java/me/golemcore/bot/port/outbound/MemorySettingsPort.java
@@ -1,0 +1,12 @@
+package me.golemcore.bot.port.outbound;
+
+/**
+ * Domain-facing access to static memory settings.
+ */
+public interface MemorySettingsPort {
+
+    MemorySettings memory();
+
+    record MemorySettings(String directory) {
+    }
+}
