+++ src/main/java/me/golemcore/bot/port/outbound/WorkspaceSettingsPort.java
@@ -1,0 +1,12 @@
+package me.golemcore.bot.port.outbound;
+
+/**
+ * Domain-facing access to workspace path settings.
+ */
+public interface WorkspaceSettingsPort {
+
+    WorkspaceSettings workspace();
+
+    record WorkspaceSettings(String filesystemWorkspace, String shellWorkspace) {
+    }
+}
