+++ src/test/java/me/golemcore/bot/support/TestPorts.java
@@ -4,33 +4,33 @@ import me.golemcore.bot.adapter.outbound.config.BotPropertiesSettingsAdapter;
 import me.golemcore.bot.adapter.outbound.i18n.MessageLocalizationAdapter;
 import me.golemcore.bot.infrastructure.config.BotProperties;
 import me.golemcore.bot.infrastructure.i18n.MessageService;
-import me.golemcore.bot.port.outbound.BotSettingsPort;
 import me.golemcore.bot.port.outbound.LocalizationPort;
+import me.golemcore.bot.port.outbound.ToolRuntimeSettingsPort;
 
 public final class TestPorts {
 
     private TestPorts() {
     }
 
-    public static BotSettingsPort settings(BotProperties properties) {
+    public static BotPropertiesSettingsAdapter settings(BotProperties properties) {
         return new BotPropertiesSettingsAdapter(properties);
     }
 
-    public static BotSettingsPort.TurnSettings turn(BotProperties.TurnProperties properties) {
+    public static ToolRuntimeSettingsPort.TurnSettings turn(BotProperties.TurnProperties properties) {
         if (properties == null) {
-            return BotSettingsPort.defaultTurnSettings();
+            return ToolRuntimeSettingsPort.defaultTurnSettings();
         }
-        return new BotSettingsPort.TurnSettings(
+        return new ToolRuntimeSettingsPort.TurnSettings(
                 properties.getMaxLlmCalls(),
                 properties.getMaxToolExecutions(),
                 properties.getDeadline());
     }
 
-    public static BotSettingsPort.ToolLoopSettings toolLoop(BotProperties.ToolLoopProperties properties) {
+    public static ToolRuntimeSettingsPort.ToolLoopSettings toolLoop(BotProperties.ToolLoopProperties properties) {
         if (properties == null) {
-            return BotSettingsPort.defaultToolLoopSettings();
+            return ToolRuntimeSettingsPort.defaultToolLoopSettings();
         }
-        return new BotSettingsPort.ToolLoopSettings(
+        return new ToolRuntimeSettingsPort.ToolLoopSettings(
                 properties.isStopOnToolFailure(),
                 properties.isStopOnConfirmationDenied(),
                 properties.isStopOnToolPolicyDenied());
